Working on 'Cancel Edit' feature
Added a cancel button and subsequent action on line 15 of `templates/components/update-form.hbs`
Added closeEdit action to lines 24-32 of `component/update-form.js` which sends the `closeEdit` action to `templates/components/update-task.hbs`
Added closeEdit action to lines 11-12 of `update-task.js` which sends the action to `templates/tasks.hbs`
Added closeEdit action to lines 26-30 of `routes/tasks.js` which switches `editing` to 'true'

Cancel action is not currently working as expected. Will be looking for a fix to the issue.
Caleb recommended not using `alias` in `update-form.js`